moved script over
